fixed proper output in test and added watching
