+++ package.json
@@ -3,7 +3,8 @@
   "name": "passport-local-simple",
   "description": "A simple implementation of a LocalStrategy for passport.",
   "scripts": {
-    "start": "node index.js",
+    "start": "nodemon index.js",
+    "test:watch": "watch-run -i -p 'index.js,app.js,./public/**/*,./test/test.js' npm run test",
     "test": "mocha"
   },
   "version": "0.0.1",
@@ -15,6 +16,8 @@
   },
   "devDependencies": {
     "mocha": "^5.2.0",
-    "request": "^2.88.0"
+    "nodemon": "^1.18.6",
+    "request": "^2.88.0",
+    "watch-run": "^1.2.5"
   }
 }
